Revert "Merge remote-tracking branch 'goog/master-chromium' into 'goog/master'" DO NOT MERGE
This reverts commit ee9208eab62e70e384912ffdc893102a880bd283, reversing
changes made to 2dad62296f6b4c462b329c48b3bcc31cfdc56acc.

+++ src/images/SkFDStream.cpp
@@ -57,7 +57,7 @@ size_t SkFDStream::read(void* buffer, size_t size) {
 #endif
                 return 0;
             }
-            return (size_t) size;
+            return size;
         } else if (NULL == buffer) {        // skip
             off_t oldCurr = ::lseek(fFD, 0, SEEK_CUR);
             if (oldCurr < 0) {
@@ -74,7 +74,7 @@ size_t SkFDStream::read(void* buffer, size_t size) {
                 return 0;   // error;
             }
             // return the actual amount we skipped
-            return (size_t) (newCurr - oldCurr);
+            return newCurr - oldCurr;
         } else {                            // read
             ssize_t actual = ::read(fFD, buffer, size);
             // our API can't return an error, so we return 0
@@ -89,3 +89,4 @@ size_t SkFDStream::read(void* buffer, size_t size) {
     }
     return 0;
 }
+
